Move instrumentation code to a separate file.
